Selective triggering based on a tag pattern
This adds the ability to specify a `tag_filter` that will only trigger
pipelines when a matching tag is pushed to the repository. It will work
with both bare and annotated tags, however it is recommended that you
use annotated tags when pushing to public repositories as they have time
and author metadata.

e.g. `tag_filter: *production` would match a semver tag of
`0.1.0-production` and trigger the pipeline.