Enable annotated tagging
This allows the resource to push annotated tags as well as lightweight
tags, meaning they have metadata attached such as the author, timestamp,
and a message. It is expected that users would use annotated tags as
part of a release process.

e.g. passing `annotate: foo/bar/annotation_file` to a resource tag
operation will result in an annotated tag being created with the
contents of the `foo/bar/annotation_file` as the annotation message.