get returns same ref requested by concourse
The in command does checkout the code at a given ref, but it does not change
it. In consequence it should report the same ref as passed from concourse
as version to checkout.

We only return a different version in the case no version is specified
by concourse (first get).

For instance when using tag_filter, the check step will return a list of valid
tags as versions. In order to keep consistency, the get command
should also return the tag if the tag_filter is defined.

We include a test to verify that get reports the same version for
different ref (hash or tag) pointing to the same commit.

Fixes #45