prevent pulling in refs from other branches
previously, merging a branch would bring all of its commit history in,
i.e. the tree that you see in `git log`.

normally, using git and other tooling, it makes for these commits to
show up in `git log`, but from a CI standpoint it doesn't really make
sense to consider those commits, as they never existed in the branch
that we're watching *in that state*. the first time these commits ever
appear in a reasonable state on the branch is the merge commit.

the fix is to just pass `--first-parent` to `git log`.

fixes #39